Add index file for exporting components
