+++ components/FAQ/questionAndAnswers.js
@@ -1,8 +1,4 @@
-import React from 'react'
-import styled from 'styled-components'
-import { Accordion } from './Accordion'
-
-const questionAndAnswers = [
+export const questionAndAnswers = [
   {
     id: 'function',
     label: 'Wie funktioniert “Hotspot or not?”',
@@ -47,36 +43,3 @@ const questionAndAnswers = [
       'Der Inzidenzwert ist die Maßzahl dafür, wie viele von 100.000 Einwohnern innerhalb von 7 Tagen an Covid-19 erkrankt sind.',
   },
 ]
-
-const Stack = styled.div`
-  > * {
-    margin-bottom: 8px;
-
-    &:last-child {
-      margin-bottom: 0;
-    }
-  }
-`
-
-export const FAQ = () => {
-  return (
-    <>
-      <h2>FAQ</h2>
-      <Stack>
-        {questionAndAnswers.map((it) => (
-          <Accordion key={it.id} itemId={it.id} label={it.label}>
-            {it.answer}
-          </Accordion>
-        ))}
-
-        <Accordion
-          itemId="support"
-          label="Ich habe ein Problem. Wo kann ich mich melden?"
-        >
-          Bei Problemen, Fragen oder Anmerkungen wenden Sie sich bitte an{' '}
-          <a href="mailto:service@hotspotornot.de">service@hotspotornot.de</a>
-        </Accordion>
-      </Stack>
-    </>
-  )
-}
