+++ icons/BarIcon.js
