+++ icons/CoronaIcon.js
