+++ icons/SocialDistancingIcon.js
