+++ pages/index.js
@@ -1,13 +1,17 @@
 import React from 'react'
 import Head from 'next/head'
 import styled from 'styled-components'
-import { Section, Main, Paragraph } from '../components/Layout'
-import Header from '../components/Header'
-import { LocationButton } from '../components/LocationButton'
-import { AddressForm } from '../components/AddressForm'
-import { CoronaIcon } from '../components/icons'
-import { FAQ } from '../components/FAQ'
-import { Footer } from '../components/Footer'
+
+import {
+  Section,
+  Main,
+  Paragraph,
+  Header,
+  LocationButton,
+  AddressForm,
+  FAQ,
+  Footer,
+} from '../components'
 
 const Actions = styled.div`
   display: flex;
@@ -53,19 +57,6 @@ export default function Home() {
     </>
   )
 }
-
-const RedCoronaIcon = styled(CoronaIcon)`
-  height: auto;
-  margin-bottom: 40px;
-  width: 125px;
-
-  & .body {
-    fill: #d73d34;
-  }
-  & .circles {
-    fill: #ad3129;
-  }
-`
 
 const I = styled.i`
   color: #d73d34;
